+++ CHANGELOG.md
@@ -6,10 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Currently, this repo is in Prerelease. When it is released, this project will adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 ========
 
-## [Unreleased]
+## [ Unreleased ]
 
 ### Changes
 
+- Input: added onChange prop to explicitly allow it to be used as a controlled component
 - Uses stricter linters and more Prettier configs.
 
 ### Breaking Changes